% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{birthday_match_prob}
\alias{birthday_match_prob}
\title{Find the probability that at least two people in a group
have the same birthday}
\usage{
birthday_match_prob(n_people)
}
\arguments{
\item{n_people}{The number of people in the group.}
}
\value{
The probability that at least
two people in a group of this size share the same birthday.
}
\description{
Find the probability that at least two people in a group
have the same birthday
}
\examples{
birthday_match_prob(30)
}
