% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbirthday_up.R
\name{facpoly}
\alias{facpoly}
\title{Function implementing the calculation of factorial polynomials
(x)_m = x! / (x-m+1)!}
\usage{
facpoly(x, m)
}
\arguments{
\item{x}{Base}

\item{m}{index}
}
\value{
(x)_m
}
\description{
Function implementing the calculation of factorial polynomials
(x)_m = x! / (x-m+1)!
}
