% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlesum.R
\name{make_tListFunc_syntax_rcpp}
\alias{make_tListFunc_syntax_rcpp}
\title{Make syntax for nested for loop function in Rcpp}
\usage{
make_tListFunc_syntax_rcpp(n)
}
\arguments{
\item{n}{The number of people in the "class"}
}
\value{
Returns a function entitled \code{compute_tList}
}
\description{
Make syntax for nested for loop function in Rcpp
}
