% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbirthday_up.R
\name{pbirthday_up}
\alias{pbirthday_up}
\title{This calculates the probability for at least one collision in a set
of n individuals sampled iid. from a vector of length N with
occurrence probabilities as given by the vector p. This is an instance
of the birthday problem with unequal occurrence probabilities.}
\source{
H\enc{ö}{oe}hle M, Happy pbirthday class of 2016, \url{http://staff.math.su.se/hoehle/blog/2017/02/13/bday.html}
}
\usage{
pbirthday_up(n, prob, method = c("R", "Rcpp", "mase1992"))
}
\arguments{
\item{n}{Size of the set}

\item{prob}{Occurrence probabilities. The length of @prob determines N.}

\item{method}{A string describing which computational method to use. "R" (the default) works in acceptable time up to n's of about 30. The "Rcpp" options works for larger n of moderate size, e.g., n=60 takes about 3 minutes.}
}
\value{
A list containing the following elements:
   \code{prob} The probability for at least one collision
   \code{tList} A matrix containing all compositions of singletons,
    dubletons, each row has the property sum(row * 1:n) == n.
   ...
}
\description{
This calculates the probability for at least one collision in a set
of n individuals sampled iid. from a vector of length N with
occurrence probabilities as given by the vector p. This is an instance
of the birthday problem with unequal occurrence probabilities.
}
\examples{
pbirthday(n=26, classes=365, coincident=2)
pbirthday_up(n=26L, prob=rep(1/365,365), method="R")$prob
pbirthday_up(n=26L, prob=rep(1/365,365), method="Rcpp")$prob
}
